Agents are now tilted
Fungus: the locationHeight variable is being set by adding a random number to the x and y coordinates of _location. This is so all the agents are not pointed in the same direction. 
addPushMovement: The check for shoving has been changed. Shoving is now done when any of the coordinates of the distance between axes is negative.
ContinuousVector:Added a checkSign(ContinuousVector vec) function to iDynoMiCS.geometry.ContinuousVector to return -1 if any of the coordinates of the vector vec is negative.